adjust support file
