+++ package.json
@@ -2,62 +2,13 @@
   "name": "cypress-angular-unit-test",
   "version": "0.0.0-development",
   "license": "MIT",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build --prod",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e",
-    "cy:open": "cypress open",
-    "cy:run": "cypress run",
-    "travis-deploy-once": "travis-deploy-once",
-    "semantic-release": "semantic-release"
-  },
+  "scripts": {},
   "private": true,
   "files": [],
   "homepage": "https://github.com/bahmutov/cypress-angular-unit-test#readme",
-  "dependencies": {
-    "@angular/animations": "7.2.15",
-    "@angular/common": "7.2.15",
-    "@angular/compiler": "7.2.15",
-    "@angular/core": "7.2.15",
-    "@angular/forms": "7.2.15",
-    "@angular/platform-browser": "7.2.15",
-    "@angular/platform-browser-dynamic": "7.2.15",
-    "@angular/router": "7.2.15",
-    "core-js": "2.6.9",
-    "rxjs": "6.5.2",
-    "tslib": "1.10.0",
-    "zone.js": "0.9.1"
-  },
+  "dependencies": {},
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.802.1",
-    "@angular/cli": "7.3.9",
-    "@angular/compiler-cli": "7.2.15",
-    "@angular/language-service": "7.2.15",
-    "@cypress/webpack-preprocessor": "4.1.0",
-    "@types/jasmine": "2.8.16",
-    "@types/jasminewd2": "2.0.6",
-    "@types/node": "8.10.51",
-    "angular2-template-loader": "0.6.2",
-    "codelyzer": "4.5.0",
-    "cypress": "4.5.0",
-    "jasmine-core": "2.99.1",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "3.1.4",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-coverage-istanbul-reporter": "2.0.5",
-    "karma-jasmine": "1.1.2",
-    "karma-jasmine-html-reporter": "0.2.2",
-    "protractor": "5.4.2",
-    "raw-loader": "0.5.1",
-    "semantic-release": "15.13.16",
-    "travis-deploy-once": "5.0.11",
-    "ts-loader": "5.4.5",
-    "ts-node": "7.0.1",
-    "tslint": "5.18.0",
-    "typescript": "3.5.2"
+    "cypress": "4.5.0"
   },
   "repository": {
     "type": "git",
